+++ changelog.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [1.1.3] - 2019-12-26
+
+- **Added** [#66](https://github.com/codyburleson/cburleson-pwa/issues/66) Pages: memories-of-momma-1 through 6
+
+## [1.1.2] - 2019-12-25
 
 - **Changed** `app-root.tsx` routing code; simplified how it handles Page Not Found cases based on the `<ion-route url=":any" component="app-404"/>` solution suggested [here](https://github.com/ionic-team/ionic/issues/18687).
+- **Changed** Google Analytics code so that it does not track when running in prod mode on localhost.
 - **Removed** unused `app-root.css` file and reference
 - **Fixed** [#49](https://github.com/codyburleson/cburleson-pwa/issues/49) Add page: access-the-was-console-from-websphere-portal-server
 - **Fixed** [#45](https://github.com/codyburleson/cburleson-pwa/issues/45) Add page: resolving-404-errors-accessing-wcm-content-through-poc-servlet